+++ src/app/app.ts
@@ -2,15 +2,14 @@ import { Component, inject } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
 import { LanguageService } from './auth/services/language.service';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-root',
-  imports: [RouterOutlet],
+  imports: [RouterOutlet, TranslateModule],
   templateUrl: './app.html',
 })
 export class App {
-  protected title = 'Task Admin';
-
   language = inject(LanguageService);
 
   constructor() {