+++ src/app/tasks/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './task-state.interface';
+export * from './mini-task-item.interface';
+export * from './task-response.interface';