+++ src/app/tasks/interfaces/task-state.interface.ts
@@ -1,0 +1,7 @@
+export enum TaskState {
+  CREATED = 'created',
+  STARTED = 'started',
+  ENDED = 'ended',
+  PAUSED = 'paused',
+  CANCELED = 'canceled',
+}