+++ src/app/tasks/services/task.service.ts
@@ -1,0 +1,23 @@
+import { HttpClient } from '@angular/common/http';
+import { inject, Injectable } from '@angular/core';
+
+import { Observable } from 'rxjs';
+import { TasksResponse, TaskState } from '../interfaces';
+import { environment } from '../../../environments/environment';
+
+const BASE_URL = environment.BASE_URL;
+
+@Injectable({
+  providedIn: 'root',
+})
+export class TaskService {
+  private http = inject(HttpClient);
+
+  getTasks(): Observable<TasksResponse> {
+    return this.http.get<TasksResponse>(`${BASE_URL}/tasks`);
+  }
+
+  getTasksByState(state: TaskState): Observable<TasksResponse> {
+    return this.http.get<TasksResponse>(`${BASE_URL}/tasks/state/${state}`);
+  }
+}